+++ CHANGELOG.md
@@ -63,6 +63,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 - [1199](https://github.com/umee-network/umee/pull/1199) Move all queries which require address input (e.g. `supplied`, `collateral_value`, `borrow_limit`) into aggregate queries `acccount_summary` or `account_balances`.
 - [1236](https://github.com/umee-network/umee/pull/1236) Add more response fields to leverage messages.
 - [1236](https://github.com/umee-network/umee/pull/1236) `Asset` -> `Coin` on leverage messages.
+- [1222](https://github.com/umee-network/umee/pull/1222) Add leverage parameter DirectLiquidationFee.
 
 ### Features
 
@@ -81,7 +82,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 - [1203](https://github.com/umee-network/umee/pull/1203) Add Swagger docs.
 - [1212](https://github.com/umee-network/umee/pull/1212) Add `util/checkers` utility package providing common check / validation functions.
 - [1220](https://github.com/umee-network/umee/pull/1220) Submit oracle prevotes / vote txs via the CLI.
-- [1236](https://github.com/umee-network/umee/pull/1236) Improve leverage event fields.
+- [1222](https://github.com/umee-network/umee/pull/1222) Liquidation reward_denom can now be either token or uToken.
 
 ### Improvements
 
@@ -91,6 +92,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 - [967](https://github.com/umee-network/umee/pull/962) Use taylor series of e^x for more accurate interest at high APY.
 - [987](https://github.com/umee-network/umee/pull/987) Streamline x/leverage CLI tests
 - [1012](https://github.com/umee-network/umee/pull/1012) Improve negative time elapsed error message
+- [1236](https://github.com/umee-network/umee/pull/1236) Improve leverage event fields.
 
 ### Bug Fixes